+++ src/detectors/builtin/inheritedStateMutation.ts
@@ -1,9 +1,13 @@
 import { InternalException } from "../../internals/exceptions";
 import { CompilationUnit } from "../../internals/ir";
-import { foldStatements } from "../../internals/tactASTUtil";
+import {
+  foldStatements,
+  collectMutations,
+  mutationNames,
+} from "../../internals/tactASTUtil";
 import { MistiTactWarning, Severity } from "../../internals/warnings";
 import { ASTDetector } from "../detector";
-import { AstStatement, isSelfId } from "@tact-lang/compiler/dist/grammar/ast";
+import { AstStatement, AstNode } from "@tact-lang/compiler/dist/grammar/ast";
 
 /**
  * An optional detector that highlights all instances where inherited trait variables
@@ -51,6 +55,11 @@ import { AstStatement, isSelfId } from "@tact-lang/compiler/dist/grammar/ast";
 export class InheritedStateMutation extends ASTDetector {
   async check(cu: CompilationUnit): Promise<MistiTactWarning[]> {
     return Array.from(cu.ast.getContracts()).reduce((acc, contract) => {
+      const contractAst = cu.ast.getContract(contract.id);
+      if (!contractAst)
+        throw InternalException.make(
+          `Cannot find contract AST: ${contract.name}`,
+        );
       const inheritedFields = cu.ast.getInheritedFields(contract.id);
       if (inheritedFields === undefined) {
         throw InternalException.make(
@@ -58,15 +67,21 @@ export class InheritedStateMutation extends ASTDetector {
         );
       }
       const inheritedFieldNames = inheritedFields.map((f) => f.name.text);
-      return acc.concat(
-        foldStatements(contract, [] as MistiTactWarning[], (acc, stmt) => {
+      const check = (node: AstNode): MistiTactWarning[] =>
+        foldStatements(node, [] as MistiTactWarning[], (acc, stmt) => {
           return this.findInheritedFieldAssignments(
             acc,
             stmt,
             inheritedFieldNames,
           );
-        }),
-      );
+        });
+      const contractWarnings = contract.declarations.reduce((acc, decl) => {
+        // Skip init functions since we cannot access self to use setters
+        return decl.kind === "function_def" || decl.kind === "receiver"
+          ? acc.concat(check(decl))
+          : acc;
+      }, [] as MistiTactWarning[]);
+      return acc.concat(contractWarnings);
     }, [] as MistiTactWarning[]);
   }
 
@@ -75,22 +90,25 @@ export class InheritedStateMutation extends ASTDetector {
     stmt: AstStatement,
     inheritedFieldNames: string[],
   ): MistiTactWarning[] {
-    if (
-      (stmt.kind === "statement_assign" ||
-        stmt.kind === "statement_augmentedassign") &&
-      stmt.path.kind === "field_access" &&
-      stmt.path.aggregate.kind === "id" &&
-      isSelfId(stmt.path.aggregate) &&
-      inheritedFieldNames.includes(stmt.path.field.text)
-    ) {
+    const intersection = <T>(l1: T[], l2: T[]): T[] =>
+      l1.filter((element) => l2.includes(element));
+    const mutations = collectMutations(stmt);
+    const foundMutations = mutations
+      ? intersection(
+          inheritedFieldNames,
+          mutationNames(mutations.mutatedFields),
+        )
+      : [];
+    if (foundMutations.length > 0) {
+      const mut = foundMutations.join(", ");
       acc.push(
         this.makeWarning(
-          "Inherited Trait Variable Mutation",
+          `Inherited trait field${foundMutations.length === 1 ? ` ${mut} is` : `s ${mut} are`} mutated`,
           Severity.LOW,
           stmt.loc,
           {
             extraDescription:
-              "Directly modifying inherited trait variables can indicate a potential error or poor design",
+              "Directly modifying inherited trait fields can indicate a potential error or poor design",
             suggestion: "Consider using setter methods to preserve invariants",
           },
         ),